==> 03_run_all_mcmc.R <==
# Run all MCMCs

# Fitted using drjacoby v1.5.3

source("04_main_fitting.R")
source("05_sensitivity_fever.R")
source("06_sensitivity_chronic.R")
source("07_sensitivity_dur.R")
source("08_sensitivity_rdt.R")
